# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "北向API接口",
        "title": "北向API接口",
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
        "/north/config_get": {
            "post": {
                "description": "获取配置接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据"
                ],
                "summary": "获取配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录凭证",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求体",
                        "name": "reqestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/constant.Request"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "type": "object"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "gzip": {
                                                            "type": "boolean"
                                                        },
                                                        "version": {
                                                            "type": "string"
                                                        }
                                                    }
                                                }
                                            ]
                                        },
                                        "version": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/constant.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "type": "object"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "nodes": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/constant.SpaceObject"
                                                            }
                                                        },
                                                        "version": {
                                                            "type": "string"
                                                        }
                                                    }
                                                }
                                            ]
                                        },
                                        "error_code": {
                                            "type": "integer"
                                        },
                                        "error_msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/north/heartbeat": {
            "post": {
                "description": "心跳接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通讯"
                ],
                "summary": "心跳",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录凭证",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/constant.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "type": "object"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "heartbeat_time": {
                                                            "type": "string"
                                                        }
                                                    }
                                                }
                                            ]
                                        },
                                        "error_code": {
                                            "type": "integer"
                                        },
                                        "error_msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/north/history_data_get": {
            "post": {
                "description": "获取历史数据接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据"
                ],
                "summary": "获取历史数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录凭证",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求体",
                        "name": "reqestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/constant.Request"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "type": "object"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "begin_time": {
                                                            "type": "integer"
                                                        },
                                                        "end_time": {
                                                            "type": "integer"
                                                        },
                                                        "point_guids": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        },
                                        "version": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/constant.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "type": "object"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "points": {
                                                            "type": "array",
                                                            "items": {
                                                                "allOf": [
                                                                    {
                                                                        "type": "object"
                                                                    },
                                                                    {
                                                                        "type": "object",
                                                                        "properties": {
                                                                            "guid": {
                                                                                "type": "string"
                                                                            },
                                                                            "tag": {
                                                                                "type": "string"
                                                                            },
                                                                            "values": {
                                                                                "type": "array",
                                                                                "items": {
                                                                                    "allOf": [
                                                                                        {
                                                                                            "type": "object"
                                                                                        },
                                                                                        {
                                                                                            "type": "object",
                                                                                            "properties": {
                                                                                                "timestamp": {
                                                                                                    "type": "integer"
                                                                                                },
                                                                                                "value": {
                                                                                                    "type": "string"
                                                                                                }
                                                                                            }
                                                                                        }
                                                                                    ]
                                                                                }
                                                                            }
                                                                        }
                                                                    }
                                                                ]
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        },
                                        "error_code": {
                                            "type": "integer"
                                        },
                                        "error_msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/north/login": {
            "post": {
                "description": "用户登录接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "reqestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/constant.Request"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "type": "object"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "password": {
                                                            "type": "string"
                                                        },
                                                        "username": {
                                                            "type": "string"
                                                        }
                                                    }
                                                }
                                            ]
                                        },
                                        "version": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/constant.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "type": "object"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "login_time": {
                                                            "type": "string"
                                                        },
                                                        "timeout": {
                                                            "type": "integer"
                                                        }
                                                    }
                                                }
                                            ]
                                        },
                                        "error_code": {
                                            "type": "integer"
                                        },
                                        "error_msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "Token": {
                                "type": "string",
                                "description": "登录凭证"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "constant.DeviceObject": {
            "type": "object",
            "properties": {
                "device_type": {
                    "type": "integer"
                },
                "guid": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "node_type": {
                    "type": "integer"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/constant.PointObject"
                    }
                },
                "path": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                }
            }
        },
        "constant.PointObject": {
            "type": "object",
            "properties": {
                "abs_value": {
                    "type": "number"
                },
                "alarm_level": {
                    "type": "integer"
                },
                "alarm_type": {
                    "type": "integer"
                },
                "ao_bound": {
                    "type": "string"
                },
                "guid": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "node_type": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "percentage": {
                    "type": "number"
                },
                "period": {
                    "type": "integer"
                },
                "status_map": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                }
            }
        },
        "constant.Request": {
            "type": "object",
            "required": [
                "data",
                "version"
            ],
            "properties": {
                "data": {
                    "type": "object"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "constant.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "error_code": {
                    "type": "integer"
                },
                "error_msg": {
                    "type": "string"
                }
            }
        },
...
        "constant.SpaceObject": {
            "type": "object",
            "properties": {
                "guid": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "node_type": {
                    "type": "integer"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/constant.DeviceObject"
                    }
                },
                "path": {
                    "type": "string"
                },
                "space_type": {
                    "type": "integer"
                },
                "tag": {
                    "type": "string"
                }
            }
        }
    }
}